backend/requirements.txt
flask
flask-cors
python-dotenv
faiss-cpu
langchain
langchain-community
langchain-text-splitters
langchain-groq